moved api back under bot_detector in bases
