--- bases/api_public/feedback/__init__.py
@@ -1,3 +1,0 @@
-from bases.api_public.feedback.routes import router
-
-__all__ = ["router"]
