--- bases/api_public/player/__init__.py
@@ -1,3 +1,0 @@
-from bases.api_public.player.routes import router
-
-__all__ = ["router"]
