--- bases/api_public/reports/__init__.py
@@ -1,3 +1,0 @@
-from bases.api_public.reports.routes import router
-
-__all__ = ["router"]
