+++ bases/bot_detector/api_public/player/__init__.py
@@ -1,0 +1,3 @@
+from bases.bot_detector.api_public.player.routes import router
+
+__all__ = ["router"]
