+++ bases/bot_detector/api_public/player/repository.py
@@ -1,8 +1,8 @@
 import logging
 
 import sqlalchemy as sqla
-from bases.api_public.player.schemas import PlayerCreate, PlayerInDB
-from bases.api_public.core._cache import SimpleALRUCache
+from bases.bot_detector.api_public.player.schemas import PlayerCreate, PlayerInDB
+from bases.bot_detector.api_public.core._cache import SimpleALRUCache
 from bot_detector.database.api_public import (
     Player as dbPlayer,
     PredictionFeedback as dbFeedback,
