+++ bases/bot_detector/api_public/reports/__init__.py
@@ -1,0 +1,3 @@
+from bases.bot_detector.api_public.reports.routes import router
+
+__all__ = ["router"]
